refactor(http): drop collection doc ids endpoint
